requirements.txt
gitpython>=3.1,<3.2
